+++ src/spentindex.h
@@ -43,7 +43,7 @@ struct CSpentIndexValue {
     int blockHeight;
     CAmount satoshis; // -1 for blinded output
     int addressType;
-    uint160 addressHash;
+    uint256 addressHash;
 
     ADD_SERIALIZE_METHODS;
 
@@ -57,7 +57,7 @@ struct CSpentIndexValue {
         READWRITE(addressHash);
     }
 
-    CSpentIndexValue(uint256 t, unsigned int i, int h, CAmount s, int type, uint160 a) {
+    CSpentIndexValue(uint256 t, unsigned int i, int h, CAmount s, int type, uint256 a) {
         txid = t;
         inputIndex = i;
         blockHeight = h;
